concept answer sort order is double

+++ packages/openchs-models/src/Concept.js
@@ -13,7 +13,7 @@ export class ConceptAnswer {
         properties: {
             uuid: 'string',
             concept: 'Concept',
-            answerOrder: 'int',
+            answerOrder: 'double',
             abnormal: 'bool',
             unique: 'bool',
             voided: {type: 'bool', default: false}
